v1.2.0 - add sorting to folder screen, set brewery as default sort
