+++ app/src/main/java/cz/tonda2/podtacky/features/coaster/presentation/list/ListViewModel.kt
@@ -2,6 +2,7 @@ package cz.tonda2.podtacky.features.coaster.presentation.list
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import cz.tonda2.podtacky.core.presentation.sortCoastersByType
 import cz.tonda2.podtacky.features.coaster.data.CoasterRepository
 import cz.tonda2.podtacky.features.coaster.domain.Coaster
 import cz.tonda2.podtacky.features.coaster.domain.CoasterSortType
@@ -12,7 +13,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
-import java.text.SimpleDateFormat
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class ListViewModel(
@@ -23,7 +23,7 @@ class ListViewModel(
         private const val TIMEOUT_MILLIS = 5_000L
     }
 
-    private val _order = MutableStateFlow(CoasterSortType.DATE)
+    private val _order = MutableStateFlow(CoasterSortType.BREWERY)
 
     val listUiState: StateFlow<ListScreenState> = _order
         .flatMapLatest { order ->
@@ -41,16 +41,6 @@ class ListViewModel(
             initialValue = ListScreenState()
         )
 
-    private fun sortCoastersByType(coasters: List<Coaster>, order: CoasterSortType): List<Coaster> {
-        val formatter = SimpleDateFormat("dd.MM.yyy")
-
-        return when (order) {
-            CoasterSortType.DATE -> coasters.sortedByDescending { formatter.parse(it.dateAdded) }
-            CoasterSortType.BREWERY -> coasters.sortedBy { it.brewery.lowercase() }
-            CoasterSortType.COUNT -> coasters.sortedByDescending { it.count }
-        }
-    }
-
     fun updateSortOrder(newOrder: CoasterSortType): Boolean {
         val currentOrder = _order.value
         if (currentOrder == newOrder) return false
@@ -66,5 +56,5 @@ class ListViewModel(
 
 data class ListScreenState(
     val coasters: List<Coaster> = emptyList(),
-    val order: CoasterSortType = CoasterSortType.DATE
+    val order: CoasterSortType = CoasterSortType.BREWERY
 )
